Rewrite performant dupes query (#9125)
## Describe your changes
Using all of the denormalized columns from the stack ending in #9121, we can drastically improve the queries for finding dupes.

A few optimizations:
- Instead of joining fingerprints against themselves, we just fetch the pretty small set of fingerprints that apply to the user in question. The user will have 6 max sh_datas for the fields we're currently searching on
- We previously loaded all matches into RAM and then did some in-memory logic to determine which ones to show. But, we needed to institute a max limit so we didn't OOM the server. So we had some weird logic before where we would load a random set of 30 matches into RAM.
- Instead, in this PR, we separate into two queries: one for matches within this tenant (for which we want more detailed information of which fingerprint kinds matched) and one for matches outside of this tenant.
- The internal matches should be small, but we add a limit of 100 matching fingerprints for safety.
- The external matches can be much more unbounded than matches inside this tenant, so we just fetch counts from the database
- These queries cold were taking less than a second on our users with the most matches, whereas before, they were taking 72 seconds

## Caveats
There's still one quirk: if there are more than 100 internal matches, we won't show them.

We also have a weird definition of "external matches": we don't consider a vault to be an external if it is also an internal match. So this causes cascading effects to the number of external matches when there are many internal matches.

For the latter, I wouldn't mind including _all_ external matches, even if they're an internal match? Of course, then for every internal match, we'll also see an external match for each tenant that the internal match onboards onto. But maybe this is actually useful? Lmk what you think. If we do this, we can just display the first N internal matches in the dashboard and say "more than N matches," making behavior very clear when there are so many matches within the current tenant

## Test Plan
I conserved the existing unit tests, just refactored a bit to use the new `Fingerprint::get_dupes` util.

Tested the SQL queries in prod, with muuuuch better performance.

Before the refactor, cold query:
```
                                                                                         QUERY PLAN

-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
--
 Nested Loop  (cost=3.53..83902.32 rows=36724 width=603) (actual time=8.041..72081.273 rows=79022 loops=1)
   ->  Nested Loop  (cost=3.11..65058.46 rows=36724 width=225) (actual time=8.012..71146.514 rows=79022 loops=1)
         Join Filter: (sv2.is_live = sv1.is_live)
         Rows Removed by Join Filter: 26
         ->  Index Scan using scoped_user_pkey on scoped_vault sv1  (cost=0.43..2.65 rows=1 width=26) (actual time=0.030..0.031 rows=1 loops=1)
               Index Cond: (id = 'su_kQnBlGX2iR3AtIw2BhamfS'::text)
         ->  Nested Loop  (cost=2.68..64405.29 rows=52042 width=250) (actual time=7.980..71117.953 rows=79048 loops=1)
         -- oof
               ->  Nested Loop  (cost=2.25..41146.42 rows=52042 width=69) (actual time=7.952..70069.333 rows=79048 loops=1)
                     Join Filter: (dl1.vault_id <> dl2.vault_id)
                     ->  Nested Loop  (cost=1.69..173.24 rows=56531 width=94) (actual time=7.916..837.365 rows=80433 loops=1)
                           ->  Nested Loop  (cost=1.12..140.56 rows=7 width=131) (actual time=7.862..10.197 rows=6 loops=1)
                                 ->  Index Scan using data_lifetime_active_by_sv_by_kind on data_lifetime dl1  (cost=0.56..29.62 rows=22 width=75) (actual time=7.830..7.868 rows=3 loops=1
)
                                       Index Cond: ((scoped_vault_id = 'su_kQnBlGX2iR3AtIw2BhamfS'::text) AND (kind = ANY ('{id.phone_number,id.email,id.ssn9}'::text[])))
                                 ->  Index Scan using fingerprint_lifetime_id on fingerprint f1  (cost=0.56..5.01 rows=3 width=106) (actual time=0.769..0.771 rows=2 loops=3)
                                       Index Cond: (lifetime_id = dl1.id)
                           ->  Index Scan using fingerprint_sh_data on fingerprint f2  (cost=0.56..3.31 rows=136 width=87) (actual time=1.637..135.247 rows=13406 loops=6)
                                 Index Cond: (sh_data = f1.sh_data)
                                 Filter: (f1.id <> id)
                                 Rows Removed by Filter: 1
                     ->  Index Scan using data_lifetime_pkey on data_lifetime dl2  (cost=0.56..0.71 rows=1 width=75) (actual time=0.859..0.859 rows=1 loops=80433)
                           Index Cond: (id = f2.lifetime_id)
                           Filter: (deactivated_seqno IS NULL)
                           Rows Removed by Filter: 0
               ->  Index Scan using scoped_user_pkey on scoped_vault sv2  (cost=0.43..0.45 rows=1 width=206) (actual time=0.011..0.011 rows=1 loops=79048)
                     Index Cond: (id = dl2.scoped_vault_id)
   ->  Index Scan using user_vault_pkey on vault v2  (cost=0.43..0.51 rows=1 width=378) (actual time=0.010..0.010 rows=1 loops=79022)
         Index Cond: (id = sv2.vault_id)
 Planning Time: 5.612 ms
 Execution Time: 72092.645 ms
(29 rows)
```

After the refactor, cold query:
```
                                                                               QUERY PLAN

-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------
-------------------------------------------------------------------------------------------------------------------------------------------------------------------------
 Index Scan using fingerprint_dupes_query on fingerprint  (cost=0.56..42.59 rows=47 width=210) (actual time=21.035..662.776 rows=25025 loops=1)
   Index Cond: ((is_live = false) AND (sh_data = ANY ('{"\\x981791397f70dd49ae6697161e3bc07e6bbcdcf9701044f461fa1fe2d2abf1e1","\\x2a6fbd270d14518da428260e435bdeedc51b331466fa95a000ce3a4f1
45923e6","\\x0d322b4a454397a2ad7ab00c0ce0b7de3b1a5899058151ceae71acfcb6de746a","\\x911d5417a537882d2886db5ea2864c57a86283baa04c844cf05df1a5a376897b","\\x3bb23d334f6b64ae36ce1cc201190b1f46
935e64a638e7d0dbef5536cc9f8d9d","\\xc0022949d8c338075249761a5fb95ee9f6793800e56ea30a3810989df56bfc2a"}'::bytea[])) AND (tenant_id = 'org_hyZP3ksCvsT0AlLqMZsgrI'::text))
   Filter: (vault_id <> 'uv_6xF1K82f7nY0oo5nZ2BMrM'::text)
   Rows Removed by Filter: 6
 Planning Time: 25.331 ms
 Execution Time: 663.946 ms
(6 rows)
```

When the query is warm, it only takes 32ms after the refactor!

+++ backend/components/api/wire_types/src/resources/dupes.rs
@@ -21,6 +21,8 @@ pub struct SameTenantDupe {
 
 #[derive(Debug, Clone, Serialize, Apiv2Schema)]
 pub struct OtherTenantDupes {
-    pub num_matches: usize, // number of distinct vaults that (1) have any sort of dupe match and (2) have not onboarded onto the same tenant as the the scoped_vault for which dupes are being queried for
-    pub num_tenants: usize, // number of distinct tenants from the vaults described above ^
+    /// Number of distinct vaults that (1) have any sort of dupe match and (2) have not onboarded onto the same tenant as the the scoped_vault for which dupes are being queried for
+    pub num_matches: i64,
+    /// Number of distinct tenants from the vaults described above ^
+    pub num_tenants: i64,
 }
